requirements.txt
Flask
python-dotenv
azure-storage-blob
gunicorn
